add jObject to project
Logic that used a parsed json hash now uses jObjects to define attributes
specs updated to reflect this change. This change was made because the hash
of slack users was being used as if it was an object rather than a data hash.
It made far more sense to make them into actual objects.